+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Features
 ### Enhancements
 - Adding aggregations in hybrid query ([#630](https://github.com/opensearch-project/neural-search/pull/630))
+- Support for post filter in hybrid query ([#633](https://github.com/opensearch-project/neural-search/pull/633))
 ### Bug Fixes
 - Fix typo for sparse encoding processor factory([#600](https://github.com/opensearch-project/neural-search/pull/600))
 - Add non-null check for queryBuilder in NeuralQueryEnricherProcessor ([#619](https://github.com/opensearch-project/neural-search/pull/619))